pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.10.4",
    "python-telegram-bot==20.7",
    "pytz>=2026.1.post1",
]
